third_party/WebKit/LayoutTests/platform/mac/http/tests/serviceworker/navigation-preload/chromium/navigation-preload-resource-timing-expected.txt

This is a testharness.js-based test.
FAIL Navigation Preload Resource Timing after terminating SW. assert_equals: The first result must contain only one entry. expected 1 but got 0
Harness: the test ran to completion.
